Moved and modified listener
Added to package directory
Added TempDir class
Added temp dir to execute actions with temp JSON file
Added initial log message for run_event_actions()
